fix: register types
